# backend/docs/openapi-3.yaml
components:
  schemas:
...
    api.ProxyStatusResponse:
      properties:
        isHealthy:
          type: boolean
        lastChecked:
          type: string
        proxyDetails:
          $ref: '#/components/schemas/api.ProxyDetailsResponse'
        proxyId:
          type: string
        responseTimeMs:
          type: integer
        status:
          type: string
      type: object
    api.ProxyTestResponse:
      properties:
        error:
          type: string
        proxyId:
          type: string
        responseTime:
          type: integer
        statusCode:
          type: integer
        success:
          type: boolean
      type: object
...
    models.CreateProxyRequest:
      properties:
        address:
          type: string
        countryCode:
          type: string
        description:
          type: string
        isEnabled:
          type: boolean
        name:
          maxLength: 255
          minLength: 1
          type: string
        notes:
          type: string
        password:
          type: string
        protocol:
          type: string
          x-enum-comments:
            ProxyProtocolHTTP: '@enum http'
            ProxyProtocolHTTPS: '@enum https'
            ProxyProtocolSOCKS4: '@enum socks4'
            ProxyProtocolSOCKS5: '@enum socks5'
          x-enum-varnames:
          - ProxyProtocolHTTP
          - ProxyProtocolHTTPS
          - ProxyProtocolSOCKS5
          - ProxyProtocolSOCKS4
        username:
          type: string
      required:
      - address
      - name
      type: object
...
    models.Proxy:
      properties:
        address:
          description: Full proxy address (e.g., 'http://user:pass@host:port')
          type: string
        city:
          $ref: '#/components/schemas/sql.NullString'
        countryCode:
          $ref: '#/components/schemas/sql.NullString'
        createdAt:
          type: string
        description:
          $ref: '#/components/schemas/sql.NullString'
        failureCount:
          $ref: '#/components/schemas/sql.NullInt32'
        host:
          $ref: '#/components/schemas/sql.NullString'
        id:
          type: string
        inputPassword:
          $ref: '#/components/schemas/sql.NullString'
        inputUsername:
          $ref: '#/components/schemas/sql.NullString'
        isEnabled:
          type: boolean
        isHealthy:
          type: boolean
        lastCheckedAt:
          $ref: '#/components/schemas/sql.NullTime'
        lastError:
          $ref: '#/components/schemas/sql.NullString'
        lastStatus:
          $ref: '#/components/schemas/sql.NullString'
        lastTested:
          $ref: '#/components/schemas/sql.NullTime'
        latencyMs:
          $ref: '#/components/schemas/sql.NullInt32'
        name:
          type: string
        notes:
          $ref: '#/components/schemas/sql.NullString'
        port:
          $ref: '#/components/schemas/sql.NullInt32'
        protocol:
          $ref: '#/components/schemas/models.ProxyProtocolEnum'
        provider:
          $ref: '#/components/schemas/sql.NullString'
        status:
          $ref: '#/components/schemas/models.ProxyStatusEnum'
        successCount:
          $ref: '#/components/schemas/sql.NullInt32'
        updatedAt:
          type: string
        username:
          $ref: '#/components/schemas/sql.NullString'
      required:
      - address
      - name
      type: object
...
    models.ProxyProtocolEnum:
      type: string
      x-enum-comments:
        ProxyProtocolHTTP: '@enum http'
        ProxyProtocolHTTPS: '@enum https'
        ProxyProtocolSOCKS4: '@enum socks4'
        ProxyProtocolSOCKS5: '@enum socks5'
      x-enum-varnames:
      - ProxyProtocolHTTP
      - ProxyProtocolHTTPS
      - ProxyProtocolSOCKS5
      - ProxyProtocolSOCKS4
...
    models.UpdateProxyRequest:
      properties:
        address:
          type: string
        countryCode:
          type: string
        description:
          type: string
        isEnabled:
          type: boolean
        name:
          maxLength: 255
          minLength: 1
          type: string
        notes:
          type: string
        password:
          type: string
        protocol:
          type: string
          x-enum-comments:
            ProxyProtocolHTTP: '@enum http'
            ProxyProtocolHTTPS: '@enum https'
            ProxyProtocolSOCKS4: '@enum socks4'
            ProxyProtocolSOCKS5: '@enum socks5'
          x-enum-varnames:
          - ProxyProtocolHTTP
          - ProxyProtocolHTTPS
          - ProxyProtocolSOCKS5
          - ProxyProtocolSOCKS4
        username:
          type: string
      type: object
...
    sql.NullInt32:
      description: Port number
      properties:
        int32:
          type: integer
        valid:
          description: Valid is true if Int32 is not NULL
          type: boolean
      type: object
    sql.NullString:
      description: Fields for input/logic, not direct DB columns if already covered
        by Address or PasswordHash
      properties:
        string:
          type: string
        valid:
          description: Valid is true if String is not NULL
          type: boolean
      type: object
    sql.NullTime:
      description: Timestamp of last health check
      properties:
        time:
          type: string
        valid:
          description: Valid is true if Time is not NULL
          type: boolean
      type: object
...
  securitySchemes:
    basic:
      scheme: basic
      type: http
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive API for domain generation, validation, and management
    operations.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Domain Flow API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /proxies:
    get:
      description: Retrieve a list of proxies with optional filtering by protocol,
        status, and health
      parameters:
      - description: Maximum number of results
        in: query
        name: limit
        schema:
          default: 100
          type: integer
      - description: Number of results to skip
        in: query
        name: offset
        schema:
          default: 0
          type: integer
      - description: Filter by protocol (http, https, socks4, socks5)
        in: query
        name: protocol
        schema:
          type: string
      - description: Filter by enabled status
        in: query
        name: isEnabled
        schema:
          type: boolean
      - description: Filter by health status
        in: query
        name: isHealthy
        schema:
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/models.Proxy'
                type: array
          description: List of proxies
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to list proxies
      summary: List proxies
      tags:
      - proxies
    post:
      description: Add a new proxy configuration
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.CreateProxyRequest'
        description: Proxy creation request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Proxy'
          description: Created proxy
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid request payload or validation failed
        "409":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Proxy with address already exists
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to create proxy
      summary: Create proxy
      tags:
      - proxies
  /proxies/{proxyId}:
    delete:
      description: Delete a proxy configuration
      parameters:
      - description: Proxy ID
        in: path
        name: proxyId
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Proxy deleted successfully
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid proxy ID format
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Proxy not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to delete proxy
      summary: Delete proxy
      tags:
      - proxies
    put:
      description: Update an existing proxy configuration
      parameters:
      - description: Proxy ID
        in: path
        name: proxyId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.UpdateProxyRequest'
        description: Proxy update request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Proxy'
          description: Updated proxy
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid request payload or validation failed
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Proxy not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to update proxy
      summary: Update proxy
      tags:
      - proxies
...
  /proxies/{proxyId}/test:
    post:
      description: Test a proxy configuration to verify it works correctly
      parameters:
      - description: Proxy ID (UUID)
        in: path
        name: proxyId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ProxyTestResponse'
          description: Proxy test results
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Proxy not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Test proxy
      tags:
      - proxies
...
  /proxies/status:
    get:
      description: Retrieve health status information for all proxies
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ProxyStatusResponse'
          description: Proxy status information
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Get proxy statuses
      tags:
      - proxies
...
servers:
- url: /api/v2